+++ views/models.ejs
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+    <title>Document</title>
+
+    <!-- BOOTSTRAP -->
+    <link href="vendor/bootstrap/css/bootstrap.min.css" rel="stylesheet">
+
+    <!-- FONTAWESOME  -->
+    <link href="vendor/fontawesome-free/css/all.min.css" rel="stylesheet" type="text/css">
+
+    <!-- CSS -->
+    <link rel="stylesheet" href="stylesheets/style.css">
+
+    <!-- FONTFAMILY -->
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap"
+        rel="stylesheet"/>
+
+</head>
+
+<body>
+    <nav>
+        <div class="nav-logo">
+            <img id="train-logo" src="/images/train.png" alt="train">
+            <h4>Ticketac</h4>
+        </div>
+        <div class="nav-links">
+            <a class="active-link" href="/">Connection </a>
+            <a href="/">My last trips </a>
+        </div>
+    </nav>
+    <main class="main-content">
+
+
+    </main>
+
+
+
+
+
+    <!-- JQUERY  -->
+    <script src="vendor/jquery/jquery.min.js"></script>
+
+    <!-- BOOSTRAP  -->
+    <script src="vendor/bootstrap/js/bootstrap.bundle.min.js"></script>
+</body>
+
+</html>
